Added NotFoundException 404 and validations in several spots in the code where we can throw this exception

+++ summative-1-jojo-yi-nara/src/main/java/com/project1/Summative1jojoyinara/controller/ConsoleController.java
@@ -1,9 +1,9 @@
 package com.project1.Summative1jojoyinara.controller;
 
+import com.project1.Summative1jojoyinara.exception.ResponseStatusException;
 import com.project1.Summative1jojoyinara.model.Console;
 import com.project1.Summative1jojoyinara.repository.ConsoleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,7 +31,7 @@ public class ConsoleController {
         if(console.isPresent()) {
             return console.get();
         }
-        return null;
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "consoleId '" + id + "' does not exist");
     }
 
     @PostMapping()
@@ -50,7 +50,12 @@ public class ConsoleController {
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void removeConsole(@PathVariable Integer id){
-        consoleRepository.deleteById(id);
+        Optional<Console> console = consoleRepository.findById(id);
+        if(console.isPresent()) {
+            consoleRepository.deleteById(id);
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "consoleId '" + id + "' does not exist");
+        }
     }
 
     @GetMapping("/manufacturer/{manufacturer}")
